Renamed and moved sidebar.json and helper.js
